pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-deep-brief"
version = "0.1.0"
description = "每日自動產出美股深度研究筆記"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]

dependencies = [
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "pydantic>=2.5.0",
    "markdown>=3.5.0",
    "PyJWT>=2.8.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
ddb = "src.app:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
